+++ R/jsPropensityGadget.R
@@ -76,6 +76,9 @@ jsPropensityGadget <- function(data, nfactor.limit = 20){
                                 uiOutput("binary_check"),
                                 uiOutput("binary_var"),
                                 uiOutput("binary_val"),
+                                uiOutput("ref_check"),
+                                uiOutput("ref_var"),
+                                uiOutput("ref_val"),
                                 uiOutput("subset_check"),
                                 uiOutput("subset_var"),
                                 uiOutput("subset_val"),
@@ -312,7 +315,11 @@ jsPropensityGadget <- function(data, nfactor.limit = 20){
 
     observeEvent(c(factor_original, input$factor_vname), {
       output$binary_check <- renderUI({
-        checkboxInput(session$ns("check_binary"), "Make binary variables")
+        checkboxInput("check_binary", "Make binary variables")
+      })
+
+      output$ref_check <- renderUI({
+        checkboxInput("check_ref", "Change reference of categorical variables")
       })
 
       output$subset_check <- renderUI({
@@ -325,7 +332,7 @@ jsPropensityGadget <- function(data, nfactor.limit = 20){
       var.conti <- setdiff(names(out), c(factor_original, input$factor_vname))
       output$binary_var <- renderUI({
         req(input$check_binary == T)
-        selectInput(session$ns("var_binary"), "Variables to dichotomize",
+        selectInput("var_binary", "Variables to dichotomize",
                     choices = var.conti, multiple = T,
                     selected = var.conti[1])
       })
@@ -337,13 +344,36 @@ jsPropensityGadget <- function(data, nfactor.limit = 20){
         for (v in seq_along(input$var_binary)){
           med <- stats::quantile(out[[input$var_binary[[v]]]], c(0.05, 0.5, 0.95), na.rm = T)
           outUI[[v]] <- splitLayout(cellWidths = c("25%", "75%"),
-                                    selectInput(session$ns(paste0("con_binary", v)), paste0("Define reference:"),
+                                    selectInput(paste0("con_binary", v), paste0("Define reference:"),
                                                 choices = c("\u2264", "\u2265", "\u003c", "\u003e"), selected = "\u2264"
                                     ),
-                                    numericInput(session$ns(paste0("cut_binary", v)), input$var_binary[[v]],
+                                    numericInput(paste0("cut_binary", v), input$var_binary[[v]],
                                                  value = med[2], min = med[1], max = med[3]
                                     )
           )
+
+        }
+        outUI
+
+      })
+    })
+
+    observeEvent(input$check_ref, {
+      var.factor <- c(factor_original, input$factor_vname)
+      output$ref_var <- renderUI({
+        req(input$check_ref == T)
+        selectInput("var_ref", "Variables to change reference",
+                    choices = var.factor, multiple = T,
+                    selected = var.factor[1])
+      })
+
+      output$ref_val <- renderUI({
+        req(input$check_ref == T)
+        req(length(input$var_ref) > 0)
+        outUI <- tagList()
+        for (v in seq_along(input$var_ref)){
+          outUI[[v]] <- selectInput(paste0("con_ref", v), paste0("Reference: ", input$var_ref[[v]]),
+                                    choices = levels(factor(out[[input$var_ref[[v]]]])), selected = levels(factor(out[[input$var_ref[[v]]]]))[2])
 
         }
         outUI
@@ -433,6 +463,20 @@ jsPropensityGadget <- function(data, nfactor.limit = 20){
             label.binary[, var_label := paste0(input$var_binary[[v]], " _group")]
             #label.binary[, val_label := paste0(c(input[[paste0("con_binary", v)]], sym.ineq[input[[paste0("con_binary", v)]]]), " ", input[[paste0("cut_binary", v)]])]
             out.label <- rbind(out.label, label.binary)
+          }
+
+        }
+      }
+
+      if (!is.null(input$check_ref)){
+        if (input$check_ref){
+          validate(
+            need(length(input$var_ref) > 0 , "No variables to change reference")
+          )
+          for (v in seq_along(input$var_ref)){
+            req(input[[paste0("con_ref", v)]])
+            out1[[input$var_ref[[v]]]] <- stats::relevel(out1[[input$var_ref[[v]]]], ref = input[[paste0("con_ref", v)]])
+            out.label[variable == input$var_ref[[v]], ':='(level = levels(out1[[input$var_ref[[v]]]]), val_label = levels(out1[[input$var_ref[[v]]]]))]
           }
 
         }
@@ -546,7 +590,7 @@ jsPropensityGadget <- function(data, nfactor.limit = 20){
 
 
 
-    mat.info <- eventReactive(c(input$indep_pscal, input$group_pscal, input$caliper), {
+    mat.info <- eventReactive(c(input$indep_pscal, input$group_pscal, input$caliper, data.info()), {
       req(input$indep_pscal)
       if (is.null(input$group_pscal) | is.null(input$indep_pscal)){
         return(NULL)